add size_t cast

+++ src/mppi_library.cpp
@@ -51,7 +51,7 @@ std::vector<double> MppiControl::run(std::vector<double> &init_state, std::vecto
     Eigen::VectorXd S_array(sampling_number_);
     Eigen::VectorXd init_state_eigen = Eigen::Map<Eigen::VectorXd>(&init_state[0], init_state.size());
 
-    for (size_t i = 0; i < this->sampling_number_; i++) {
+    for (size_t i = 0; i < (size_t)this->sampling_number_; i++) {
         input_array[i] = this->generate_input_array(mu, sigma);
         state_array[i] = this->generate_model_state(
             input_array[i],
@@ -71,7 +71,7 @@ std::vector<double> MppiControl::run(std::vector<double> &init_state, std::vecto
 
     Eigen::VectorXd input(input_dim_);
     input.setZero();
-    for (size_t i = 0; i < sampling_number_; i++) input += weight_normal(i) * input_array[i].row(0).transpose();
+    for (size_t i = 0; i < (size_t)sampling_number_; i++) input += weight_normal(i) * input_array[i].row(0).transpose();
 
     std::vector<double> return_input(input.data(), input.data() + input.size());
     return return_input;
@@ -114,7 +114,7 @@ Eigen::MatrixXd MppiControl::generate_model_state(const Eigen::MatrixXd &input_a
 {
     Eigen::MatrixXd state_array(this->predict_horizon_, this->input_dim_);
     state_array.row(0) = init_state;
-    for (size_t i = 1; i < predict_horizon_; i++) state_array.row(i) = this->model(input_array.row(i), state_array.row(i-1));
+    for (size_t i = 1; i < (size_t)predict_horizon_; i++) state_array.row(i) = this->model(input_array.row(i), state_array.row(i-1));
     return state_array;
 }
 Eigen::VectorXd MppiControl::model(const Eigen::VectorXd &input, const Eigen::VectorXd &pre_state)
@@ -153,10 +153,10 @@ Eigen::MatrixXd MppiControl::generate_input_array(Eigen::VectorXd mu, Eigen::Mat
 
     std::random_device rd;
     std::mt19937 gen(rd());
-    for (size_t i = 0; i < this->predict_horizon_; i++)
+    for (size_t i = 0; i < (size_t)this->predict_horizon_; i++)
     {
         input_array.row(i) = this->sample_multivariate_normal(mu, sigma, gen);
-        for (size_t j = 0; j < this->input_dim_; j++) input_array(i, j) = this->clamp(input_array(i, j), max_input_value_[j], -max_input_value_[j]);
+        for (size_t j = 0; j < (size_t)this->input_dim_; j++) input_array(i, j) = this->clamp(input_array(i, j), max_input_value_[j], -max_input_value_[j]);
     }
     return input_array;
 }
